moved processing module names back
