+++ nwbinspector/checks/general.py
@@ -1,10 +1,7 @@
 """Check functions that examine any general neurodata_type with the available attrbutes."""
-from pynwb import ProcessingModule
-
 from ..register_checks import register_check, InspectorMessage, Importance
 
 COMMON_DESCRIPTION_PLACEHOLDERS = ["no description", "no desc", "none"]
-PROCESSING_MODULE_CONFIG = ["ophys", "ecephys", "icephys", "behavior", "misc", "ogen", "retinotopy"]
 
 
 @register_check(importance=Importance.CRITICAL, neurodata_type=None)
@@ -24,13 +21,3 @@ def check_description(obj):
         return InspectorMessage(message="Description is missing.")
     if obj.description.lower().strip(".") in common_description_placeholders:
         return InspectorMessage(message="Description is a placeholder.")
-
-
-@register_check(importance=Importance.BEST_PRACTICE_SUGGESTION, neurodata_type=ProcessingModule)
-def check_processing_module_name(processing_module: ProcessingModule):
-    """Check if the name of a processing module is of a valid modality."""
-    if processing_module.name not in PROCESSING_MODULE_CONFIG:
-        return InspectorMessage(
-            f"Processing module is named {processing_module.name}. It is recommended to use the "
-            f"schema module names: {', '.join(PROCESSING_MODULE_CONFIG)}"
-        )
