+++ nwbinspector/checks/nwbfile_metadata.py
@@ -1,7 +1,7 @@
 """Check functions that examine general NWBFile metadata."""
 import re
 
-from pynwb import NWBFile
+from pynwb import NWBFile, ProcessingModule
 from pynwb.file import Subject
 
 from ..register_checks import register_check, InspectorMessage, Importance
@@ -11,6 +11,8 @@ duration_regex = (
     r"?M)?(\d+(?:\.\d+)?S)?)?$"
 )
 species_regex = r"[A-Z][a-z]* [a-z]+"
+
+PROCESSING_MODULE_CONFIG = ["ophys", "ecephys", "icephys", "behavior", "misc", "ogen", "retinotopy"]
 
 
 @register_check(importance=Importance.BEST_PRACTICE_SUGGESTION, neurodata_type=NWBFile)
@@ -80,6 +82,16 @@ def check_subject_species(subject: Subject):
     if not re.fullmatch(species_regex, subject.species):
         return InspectorMessage(
             message="Species should be in latin binomial form, e.g. 'Mus musculus' and 'Homo sapiens'",
+        )
+
+
+@register_check(importance=Importance.BEST_PRACTICE_SUGGESTION, neurodata_type=ProcessingModule)
+def check_processing_module_name(processing_module: ProcessingModule):
+    """Check if the name of a processing module is of a valid modality."""
+    if processing_module.name not in PROCESSING_MODULE_CONFIG:
+        return InspectorMessage(
+            f"Processing module is named {processing_module.name}. It is recommended to use the "
+            f"schema module names: {', '.join(PROCESSING_MODULE_CONFIG)}"
         )
